+++ mlflow_cratedb/patch/mlflow/model.py
@@ -1,5 +1,5 @@
 import sqlalchemy as sa
-from mlflow.store.tracking.dbmodels.models import SqlEvaluationDatasetRecord
+from mlflow.store.tracking.dbmodels.models import SqlEvaluationDatasetRecord, SqlGatewayEndpoint
 from sqlalchemy.event import listen
 from sqlalchemy_cratedb.support import check_uniqueness_factory
 
@@ -19,6 +19,7 @@ def polyfill_uniqueness_constraints():
     from mlflow.store.model_registry.dbmodels.models import SqlRegisteredModel
     from mlflow.store.tracking.dbmodels.models import SqlExperiment
 
+    listen(SqlGatewayEndpoint, "before_insert", check_uniqueness_factory(SqlGatewayEndpoint, "workspace", "name"))
     listen(
         SqlEvaluationDatasetRecord,
         "before_insert",