+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 
 ## in progress
+- Documentation: Added setup for Docker Compose or Podman Compose
 
 ## 2026-03-30 v3.10.2
 - Amalgamation: Started supporting huey jobs that are invoked in subprocesses