+++ CHANGES.md
@@ -3,8 +3,10 @@
 
 ## in progress
 - Documentation: Added setup for Docker Compose or Podman Compose
-- Model: Updated tables `trace_info`, `trace_metrics`, `trace_tags`, `trace_request_metadata`
-- Model: Added tables `assessments`, `spans`, `span_metrics`
+- Model: Updated tables `trace_info`, `trace_metrics`, `trace_tags`,
+  `trace_request_metadata`
+- Model: Added tables `assessments`, `endpoints`, `endpoint_bindings`,
+  `endpoint_model_mappings`, `endpoint_tags`, `spans`, `span_metrics`
 
 ## 2026-03-30 v3.10.2
 - Amalgamation: Started supporting huey jobs that are invoked in subprocesses