+++ CHANGES.md
@@ -7,6 +7,7 @@
   `trace_request_metadata`
 - Model: Added tables `assessments`, `endpoints`, `endpoint_bindings`,
   `endpoint_model_mappings`, `endpoint_tags`, `spans`, `span_metrics`
+- Model: Renamed field names that contain dots
 
 ## 2026-03-30 v3.10.2
 - Amalgamation: Started supporting huey jobs that are invoked in subprocesses